# Conflicts:
#	app/build.gradle
#	app/src/pitstop/java/com/pitstop/ui/add_car/AddCarPresenter.java

+++ app/src/pitstop/java/com/pitstop/database/LocalPidAdapter.java
@@ -71,7 +71,7 @@ public class LocalPidAdapter {
                 pidData.setDataNumber(c.getString(c.getColumnIndex(TABLES.PID.KEY_DATANUM)));
                 pidData.setRtcTime(c.getString(c.getColumnIndex(TABLES.PID.KEY_RTCTIME)));
                 pidData.setTimeStamp(c.getString(c.getColumnIndex(TABLES.PID.KEY_TIMESTAMP)));
-                pidData.setTripId(c.getInt(c.getColumnIndex(TABLES.PID.KEY_TRIP_ID)));
+                pidData.setTripId(c.getLong(c.getColumnIndex(TABLES.PID.KEY_TRIP_ID)));
                 pidData.setPids(c.getString(c.getColumnIndex(TABLES.PID.KEY_PIDS)));
                 pidData.setMileage(c.getDouble(c.getColumnIndex(TABLES.PID.KEY_MILEAGE)));
                 pidData.setCalculatedMileage(c.getDouble(c.getColumnIndex(TABLES.PID.KEY_CALCULATED_MILEAGE)));
